requirements.txt
torch
torchvision
numpy>=1.19.2
opencv-python>=4.5.1
torch>=1.7.1
torchvision>=0.8.2
matplotlib>=3.3.4
grpcio
protobuf
psutil
pynvml